refactor: Rename project to `coral`
