+++ src/coral/whisper.py
